requirements.txt
fastapi
uvicorn[standard]
websockets
httpx
openai>=1.50.0
python-dotenv
pydantic
audioop-lts
